calculate xvp status clientside
