+++ kit/dapp/src/lib/queries/xvp/xvp-calculated.ts
@@ -2,7 +2,6 @@ import "server-only";
 
 import type { CurrencyCode } from "@/lib/db/schema-settings";
 import { safeParse } from "@/lib/utils/typebox";
-import { isBefore } from "date-fns";
 import { getAddress } from "viem";
 import { getAssetsPricesInUserCurrency } from "../asset-price/asset-price";
 import {
@@ -10,46 +9,7 @@ import {
   XvPSettlementSchema,
   type OnChainXvPSettlement,
   type XvPSettlement,
-  type XvPStatus,
 } from "./xvp-schema";
-
-/**
- * Calculates the status of an XvP settlement based on execution, cancellation, expiration, and approvals
- *
- * @param settlement - The raw settlement data from The Graph
- * @returns The calculated XvP status
- */
-function calculateXvPStatus(settlement: OnChainXvPSettlement): XvPStatus {
-  if (settlement.executed) {
-    return "EXECUTED";
-  }
-
-  if (settlement.cancelled) {
-    return "CANCELLED";
-  }
-
-  const isExpired = isBefore(settlement.cutoffDate, new Date());
-  if (isExpired) {
-    return "EXPIRED";
-  }
-
-  // Ensure flows is an array, default to empty if not present
-  const flows = Array.isArray(settlement.flows) ? settlement.flows : [];
-
-  const approvalsRequiredCount = flows.length;
-  const actualApprovalsCount = settlement.approvals.filter(
-    (approval) => approval.approved
-  ).length;
-
-  if (
-    approvalsRequiredCount > 0 &&
-    actualApprovalsCount === approvalsRequiredCount
-  ) {
-    return "READY";
-  }
-
-  return "PENDING";
-}
 
 /**
  * Enriches a single XvP settlement with calculated fields like assetPrice for flows
@@ -88,8 +48,6 @@ export async function calculateXvPSettlement(
     });
   });
 
-  const status = calculateXvPStatus(onChainSettlement);
-
   return safeParse(XvPSettlementSchema, {
     ...onChainSettlement,
     flows: calculatedFlows,
@@ -97,6 +55,5 @@ export async function calculateXvPSettlement(
       amount: totalPriceAmount,
       currency: userCurrency,
     },
-    status,
   });
 }
